backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==4.1.2
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
pytz==2022.5
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==4.1.2
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
pytz==2022.5
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
